Melhorias

+++ Projeto.Edson.DTO/Output/TemperaturaDTO.cs
@@ -5,7 +5,7 @@ namespace Projeto.Edson.DTO.Output
     public class TemperaturaDTO
     {
         public string Cidade { get; set; }
-        public double Temperatura { get; set; }
+        public decimal Temperatura { get; set; }
         public DateTime DataHora { get; set; }
     }
 }
